+++ src/ts/Ads/Views/OpenMeasurement/AdmobOpenMeasurementController.ts
@@ -15,6 +15,7 @@ import { OpenMeasurementUtilities } from 'Ads/Views/OpenMeasurement/OpenMeasurem
 import { ThirdPartyEventManager, ThirdPartyEventMacro } from 'Ads/Managers/ThirdPartyEventManager';
 import { SDKMetrics, AdmobMetric } from 'Ads/Utilities/SDKMetrics';
 import { Campaign } from 'Ads/Models/Campaign';
+import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 
 export class AdmobOpenMeasurementController extends OpenMeasurementController {
 
@@ -78,7 +79,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
         verificationResources.forEach((resource) => {
             const om = new OpenMeasurement<AdMobCampaign>(this._platform, this._core, this._clientInfo, this._campaign, this._placement, this._deviceInfo, this._thirdPartyEventManager, resource.vendorKey);
             this.setupOMInstance(om, resource);
-            if (OpenMeasurementUtilities.isDoubleClickGoogle(resource.vendorKey)) {
+            if (CustomFeatures.isDoubleClickGoogle(resource.vendorKey)) {
                 SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickInstanceCreated, {
                     'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(resource.vendorKey)
                 });
@@ -172,7 +173,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
 
             this._omInstances.forEach((om) => {
                 const verificationresource = om.getVerificationResource();
-                if (OpenMeasurementUtilities.isDoubleClickGoogle(verificationresource.vendorKey)) {
+                if (CustomFeatures.isDoubleClickGoogle(verificationresource.vendorKey)) {
                     SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions, {
                         'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
                     });
@@ -189,7 +190,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
 
             this._omInstances.forEach((om) => {
                 const verificationresource = om.getVerificationResource();
-                if (OpenMeasurementUtilities.isDoubleClickGoogle((verificationresource.vendorKey))) {
+                if (CustomFeatures.isDoubleClickGoogle((verificationresource.vendorKey))) {
                     SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMImpressions, {
                         'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
                     });
@@ -213,7 +214,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             event.data.vendorkey = verificationresource.vendorKey;
             om.sessionStart(event);
 
-            if (OpenMeasurementUtilities.isDoubleClickGoogle(verificationresource.vendorKey)) {
+            if (CustomFeatures.isDoubleClickGoogle(verificationresource.vendorKey)) {
                 SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMStarts, {
                     'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(verificationresource.vendorKey)
                 });