+++ src/ts/Ads/Views/OpenMeasurement/OpenMeasurement.ts
@@ -439,7 +439,7 @@ export class OpenMeasurement<T extends Campaign> extends View<T> {
                 SDKMetrics.reportMetricEvent(OMMetric.IASVerificatonInjected);
             }
 
-            if (OpenMeasurementUtilities.isDoubleClickGoogle(vendorKey) && this._campaign instanceof AdMobCampaign) {
+            if (CustomFeatures.isDoubleClickGoogle(vendorKey) && this._campaign instanceof AdMobCampaign) {
                 SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMInjections, {
                     'dckey': OpenMeasurementUtilities.getDcKeyMetricTag(vendorKey)
                 });