+++ src/ts/Core/Models/__mocks__/ClientInfo.ts
@@ -2,7 +2,7 @@ import { ClientInfo as Base } from 'Core/Models/ClientInfo';
 
 export type ClientInfoMock = Base & {
     getSdkVersionName: jest.Mock;
-    getApplicationName: jest.Mock
+    getApplicationName: jest.Mock;
     getTestMode: jest.Mock;
     getSdkVersion: jest.Mock;
     getGameId: jest.Mock;