+++ src/ts/VAST/AdUnits/VastAdUnitParametersFactory.ts
@@ -3,9 +3,9 @@ import { IVastAdUnitParameters } from 'VAST/AdUnits/VastAdUnit';
 import { VastCampaign } from 'VAST/Models/VastCampaign';
 import { IAdUnitParameters } from 'Ads/AdUnits/AbstractAdUnit';
 import { VastVideoOverlay } from 'Ads/Views/VastVideoOverlay';
-import { VastEndScreen, IVastEndscreenParameters } from 'VAST/Views/VastEndScreen';
 import { VastAdVerification } from 'VAST/Models/VastAdVerification';
 import { VastOpenMeasurementFactory } from 'Ads/Views/OpenMeasurement/VastOpenMeasurementFactory';
+import { VastStaticEndScreen } from 'VAST/Views/VastStaticEndScreen';
 
 export class VastAdUnitParametersFactory extends AbstractAdUnitParametersFactory<VastCampaign, IVastAdUnitParameters> {
     protected createParameters(baseParams: IAdUnitParameters<VastCampaign>) {
@@ -17,7 +17,6 @@ export class VastAdUnitParametersFactory extends AbstractAdUnitParametersFactory
         }
 
         const overlay = new VastVideoOverlay(baseParams, baseParams.privacy, this.showGDPRBanner(baseParams), showPrivacyDuringVideo);
-        let vastEndScreen: VastEndScreen | undefined;
 
         const vastAdUnitParameters: IVastAdUnitParameters = {
             ... baseParams,
@@ -25,16 +24,8 @@ export class VastAdUnitParametersFactory extends AbstractAdUnitParametersFactory
             overlay: overlay
         };
 
-        if (baseParams.campaign.hasStaticEndscreen() || baseParams.campaign.hasIframeEndscreen() || baseParams.campaign.hasHtmlEndscreen()) {
-            const vastEndscreenParameters: IVastEndscreenParameters = {
-                campaign: baseParams.campaign,
-                clientInfo: baseParams.clientInfo,
-                country: baseParams.coreConfig.getCountry(),
-                hidePrivacy: baseParams.adsConfig.getHidePrivacy()
-            };
-
-            vastEndScreen = new VastEndScreen(baseParams.platform, vastEndscreenParameters, baseParams.privacy);
-            vastAdUnitParameters.endScreen = vastEndScreen;
+        if (baseParams.campaign.hasStaticEndscreen()) {
+            vastAdUnitParameters.endScreen = new VastStaticEndScreen(baseParams);
         }
 
         const adVerifications: VastAdVerification[] = baseParams.campaign.getVast().getAdVerifications();